Move attributes to their own files.
