+++ Weaver/WeaverTests/Realm.FakeForWeaverTests/Realm.FakeForWeaverTests.csproj
@@ -38,12 +38,6 @@
     <!-- A reference to the entire .NET Framework is automatically included -->
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes.cs">
-      <Link>Attributes.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\Shared\Realm.Shared\PreserveAttribute.cs">
-      <Link>PreserveAttribute.cs</Link>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RealmList.cs" />
     <Compile Include="RealmObject.cs" />
@@ -52,9 +46,34 @@
     </Compile>
     <Compile Include="Exceptions\RealmException.cs" />
     <Compile Include="Realm.cs" />
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\Attributes.cs">
+      <Link>Attributes\Attributes.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\BacklinkAttribute.cs">
+      <Link>Attributes\BacklinkAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\IgnoredAttribute.cs">
+      <Link>Attributes\IgnoredAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\IndexedAttribute.cs">
+      <Link>Attributes\IndexedAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\MapToAttribute.cs">
+      <Link>Attributes\MapToAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\PreserveAttribute.cs">
+      <Link>Attributes\PreserveAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\PrimaryKeyAttribute.cs">
+      <Link>Attributes\PrimaryKeyAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\RequiredAttribute.cs">
+      <Link>Attributes\RequiredAttribute.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Exceptions\" />
+    <Folder Include="Attributes\" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <Import Project="$(SolutionDir)packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets" Condition="Exists('$(SolutionDir)packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets')" />
