+++ Weaver/WeaverTests/RealmWeaver.Tests/RealmWeaver.Tests.csproj
@@ -73,6 +73,9 @@
   <ItemGroup>
     <Compile Include="WeaverTests.cs" />
     <Compile Include="Verifier.cs" />
+    <Compile Include="..\..\..\Shared\Realm.Shared\Attributes\Attributes.cs">
+      <Link>Attributes\Attributes.cs</Link>
+    </Compile>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\..\..\packages\Fody.1.29.4\build\portable-net+sl+win+wpa+wp\Fody.targets" Condition="Exists('..\..\..\packages\Fody.1.29.4\build\portable-net+sl+win+wpa+wp\Fody.targets')" />
@@ -101,6 +104,9 @@
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Attributes\" />
   </ItemGroup>
   <Import Project="..\..\..\packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets" Condition="Exists('..\..\..\packages\StyleCop.MSBuild.4.7.54.0\build\StyleCop.MSBuild.Targets')" />
 </Project>
